Replace CustomActions::TableCell by ViewComponent
https://community.openproject.org/wp/47008
